This Shit Dont Work !

+++ build.sh
@@ -1,6 +1,13 @@
 #! /usr/bin/sh
 
+echo -- Building!
+
 mkdir dist -p
 
 g++ $@ -pthread -Ivendor/include/ -o dist/bot src/*.cc -Lvendor/lib/curlpp -lssl -lcrypto -lcurl -lcurlpp  -Wl,-rpath='$ORIGIN'
-cp vendor/lib/**/*.so* dist
+
+if test -e dist; then
+    cp vendor/lib/**/*.so* dist
+fi
+
+echo -- Done!
